Add social media buttons for the blog (#324)
And added Rickie's recording from Wednesday.
