+++ .codeclimate.yml
@@ -1,0 +1,3 @@
+version: "2"
+exclude_patterns:
+  - "src/utils/api.tsx"
