+++ src/content/justice-for-rickie-slaughter.mdx
@@ -9,6 +9,9 @@ tags: pro-bono,black-lives-matter,criminal-justice
 import { Image } from '@components/image';
 import { VegasDirections } from '@components/vegasDirections';
 
+> Rickie shared an update on July 1, 2020, thanking all of you for checking
+> out this website. [Listen to it here](/audio/20200701-rickie-update.m4a).
+
 Rickie Lamont Slaughter is an innocent man who is serving a prison sentence
 for a crime he did not commit.
