+++ src/content/slaughter-v-escamilla.mdx
@@ -1,7 +1,8 @@
 ---
 title: Slaughter v. Escamilla
 slug: /blog/slaughter-v-escamilla
-updatedAt: "2019-07-01"
+featuredImage: rickie.jpg
+updatedAt: "2020-07-01"
 tags: criminal-justice,racial-justice
 ---
